Move features into features folder
