+++ features/fish_market.lua
@@ -19,7 +19,7 @@ make pet follow you moar
 local Event = require 'utils.event'
 local Token = require 'utils.global_token'
 local Task = require 'utils.Task'
-local PlayerStats = require 'player_stats'
+local PlayerStats = require 'features.player_stats'
 local Game = require 'utils.game'
 local Utils = require 'utils.utils'
