+++ features/wood_market.lua
@@ -1,7 +1,7 @@
 local Event = require 'utils.event'
 local Token = require 'utils.global_token'
 local Task = require 'utils.Task'
-local PlayerStats = require 'player_stats'
+local PlayerStats = require 'features.player_stats'
 local Game = require 'utils.game'
 local Utils = require "utils.utils"
