+++ map_gen/combined/cave_miner/cave_miner.lua
@@ -21,7 +21,7 @@ local math = require 'utils.math'
 local market_items = require "map_gen.combined.cave_miner.market_items"
 local Game = require 'utils.game'
 
-local Info = require 'info'
+local Info = require 'features.gui.info'
 
 Info.set_map_name('Cave Miner')
 Info.set_map_description([[
