+++ map_gen/presets/crash_site/outpost_builder.lua
@@ -4,7 +4,7 @@ local Global = require 'utils.global'
 local Event = require 'utils.event'
 local Task = require 'utils.Task'
 local Market = require 'map_gen.presets.crash_site.market'
-local PlayerStats = require 'player_stats'
+local PlayerStats = require 'features.player_stats'
 
 local b = require 'map_gen.shared.builders'
